+++ Net_Core/Web.Api/Program.cs
@@ -1,6 +1,7 @@
 using GrowthWare.Framework;
 using GrowthWare.Web.Support.Jwt;
 using GrowthWare.Web.Support.Utilities;
+using GrowthWare.Web.Support.Helpers;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
@@ -94,7 +95,7 @@ IHttpContextAccessor? httpContextAccessor = app.Services.GetService<IHttpContext
 if (httpContextAccessor != null)
 {
     SecurityEntityUtility.SetHttpContextAccessor(httpContextAccessor);
-    SessionController.SetHttpContextAccessor(httpContextAccessor);
+    SessionHelper.SetHttpContextAccessor(httpContextAccessor);
 }
 
 if (app.Environment.IsDevelopment())